move IndexValue to builtins.py
