+++ numba/parfor.py
@@ -50,7 +50,6 @@ from numba.ir_utils import (
     simplify_CFG,
     has_no_side_effect,
     canonicalize_array_math,
-    IndexValue,
     find_callname,
     guard,
     require,
@@ -62,7 +61,6 @@ from numba.ir_utils import (
 from numba.analysis import (compute_use_defs, compute_live_map,
                             compute_dead_maps, compute_cfg_from_blocks)
 from numba.controlflow import CFGraph
-from numba.targets.builtins import get_type_max_value, get_type_min_value
 from numba.typing import npydecl, signature
 from numba.types.functions import Function
 from numba.array_analysis import (random_int_args, random_1arg_size,
@@ -91,17 +89,17 @@ _np_reduce_calls = ['min', 'max']
 
 def argmin_parallel_impl(A):
     init_val = numba.targets.builtins.get_type_max_value(A.dtype)
-    ival = numba.ir_utils.IndexValue(0, init_val)
+    ival = numba.typing.builtins.IndexValue(0, init_val)
     for i in numba.prange(len(A)):
-        curr_ival = numba.ir_utils.IndexValue(i, A[i])
+        curr_ival = numba.typing.builtins.IndexValue(i, A[i])
         ival = min(ival, curr_ival)
     return ival.index
 
 def argmax_parallel_impl(A):
     init_val = numba.targets.builtins.get_type_min_value(A.dtype)
-    ival = numba.ir_utils.IndexValue(0, init_val)
+    ival = numba.typing.builtins.IndexValue(0, init_val)
     for i in numba.prange(len(A)):
-        curr_ival = numba.ir_utils.IndexValue(i, A[i])
+        curr_ival = numba.typing.builtins.IndexValue(i, A[i])
         ival = max(ival, curr_ival)
     return ival.index
 
@@ -736,6 +734,7 @@ class ParforPass(object):
         return parfor
 
     def _reduction_to_parfor(self, equiv_set, lhs, expr):
+        from numba.targets.builtins import get_type_max_value, get_type_min_value
         call_name, mod_name = find_callname(self.func_ir, expr)
         args = expr.args
         kws = dict(expr.kws)
